# app/api/stories/[id]/publish/route.ts
import { NextResponse } from "next/server";
import { getCurrentUserId } from "@/lib/auth";
import { scanTextBundle, storyPublishTextParts } from "@/lib/content-filter";
import { getDb, nowIso } from "@/lib/db";
import { createNotification } from "@/lib/notifications";
import { enqueueSensitivePublishBlock } from "@/lib/moderation-queue";
import { getRequestIp, rateLimitAllow } from "@/lib/rate-limit";

export async function POST(
  req: Request,
  { params }: { params: Promise<{ id: string }> },
) {
  const { id } = await params;
  const userId = await getCurrentUserId();
  const db = await getDb();

  const rlUser = rateLimitAllow(`publish:${userId}`, 35, 3_600_000);
  if (!rlUser.ok) {
    return NextResponse.json(
      { code: 429, msg: "发布过于频繁，请稍后再试" },
      {
        status: 429,
        headers: { "Retry-After": String(Math.ceil(rlUser.retryAfterMs / 1000)) },
      },
    );
  }
  const rlIp = rateLimitAllow(`publish_ip:${getRequestIp(req)}`, 150, 3_600_000);
  if (!rlIp.ok) {
    return NextResponse.json(
      { code: 429, msg: "当前网络发布过于频繁，请稍后再试" },
      {
        status: 429,
        headers: { "Retry-After": String(Math.ceil(rlIp.retryAfterMs / 1000)) },
      },
    );
  }

  const story = await db.get<{ id: string; title: string }>(
    "SELECT id, title FROM stories WHERE id = ? AND author_id = ?",
    id,
    userId,
  );
  if (!story) {
    return NextResponse.json({ code: 404, msg: "故事不存在" }, { status: 404 });
  }

  const parts = await storyPublishTextParts(db, id);
  const scan = scanTextBundle(parts);
  if (!scan.ok) {
    try {
      await enqueueSensitivePublishBlock(db, {
        contentType: "story",
        targetId: id,
        submitterUserId: userId,
      });
    } catch (e) {
      console.error("[moderation enqueue]", e);
    }
    return NextResponse.json({ code: 400, msg: scan.msg }, { status: 400 });
  }

  const now = nowIso();
  await db.run(
    "UPDATE stories SET status = 'published', publish_at = ?, updated_at = ? WHERE id = ?",
    now,
    now,
    id,
  );

  const followers = await db.all<{ user_id: string }[]>(
    "SELECT user_id FROM follows WHERE author_id = ?",
    userId,
  );
  for (const row of followers) {
    await createNotification(db, row.user_id, "author_update", {
      author_id: userId,
      story_id: id,
      story_title: story.title,
    });
  }

  return NextResponse.json({ code: 200, msg: "发布成功", data: { publish_at: now } });
}