# docs/openapi.yaml
openapi: 3.0.3
info:
  title: StoryForge MVP API
  version: 0.3.0
  description: MVP API — avatar upload (JPG/PNG ≤5MB, thumb 200²), Google OAuth (JWT sf_session), admin moderation, chat SSE (heartbeat + timeout), snapshots, stories, branches export annotate, profile, account deletion (§2.3), moderation, export, feed (author_display for deleted users). Production mutations on /api/* Origin-guard (middleware). Admins STORYFORGE_ADMIN_USER_IDS.
servers:
  - url: http://localhost:3000
paths:
...
  /api/stories/{id}/publish:
    post:
      summary: Publish story
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
...
  /api/stories/{id}/outline:
    get:
      summary: List outline nodes (author only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
    post:
      summary: Create outline node
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
  /api/stories/{id}/outline/{nodeId}:
    patch:
      summary: Update outline node or reorder (move up/down)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: nodeId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
    delete:
      summary: Delete outline subtree
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: nodeId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
...
  /api/stories/{id}/export:
    post:
      summary: Export story (markdown / txt / pdf / epub)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [format]
              properties:
                format:
                  type: string
                  enum: [markdown, txt, pdf, epub]
                branch_mode:
                  type: string
                  enum: [none, annotate]
                  description: Append story_branches appendix with [分支] lines
      responses:
        "200":
          description: File download or markdown fallback
...